+++ sdk/identity/azure-identity/azure/identity/aio/__init__.py
@@ -7,9 +7,7 @@ from .credentials import (
     AsyncCertificateCredential,
     AsyncClientSecretCredential,
     AsyncEnvironmentCredential,
-    AsyncImdsCredential,
     AsyncManagedIdentityCredential,
-    AsyncMsiCredential,
     AsyncTokenCredentialChain,
 )
 
@@ -26,8 +24,6 @@ __all__ = [
     "AsyncClientSecretCredential",
     "AsyncDefaultAzureCredential",
     "AsyncEnvironmentCredential",
-    "AsyncImdsCredential",
     "AsyncManagedIdentityCredential",
-    "AsyncMsiCredential",
     "AsyncTokenCredentialChain",
 ]
